+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+- Nothing
+
+## 0.0.1 - 2018-02-26
+
+- Echo service implementation
 - Creation from [gomeet](https://github.com/gomeet/gomeet)
 
 ```shell
 $ gomeet new github.com/gomeet-examples/svc-echo \
-  --default-prefixes=svc-,gomeet-svc- \
+  --default-prefixes=gomeet-svc-,svc- \
   --proto-name=pb
 ```